+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Upcoming
+* Fix: Instability on pressing the connect and disconnect button
+
 # v6.0.676
 * Fix: Improve security by blocking multicast traffic
 * Fix: No internet connection on some devices